+++ android/app/src/main/java/com/finly/app/MainActivity.java
@@ -13,6 +13,7 @@ import com.getcapacitor.BridgeActivity;
 public class MainActivity extends BridgeActivity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
+        registerPlugin(FinlyBiometricPlugin.class);
         super.onCreate(savedInstanceState);
         setupCutoutAndFullScreen();
     }